added description and bg
